\name{aisoph-package}
\alias{aisoph-package}
\docType{package}
\title{Additive Isotonic Proportional Hazards Model}
\description{Nonparametric estimation of additive isotonic covariate effects for proportional hazards model.}
\details{
\tabular{ll}{
Package: \tab aisoph\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2023-03-03\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [aut, cre]
Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yunro Chung, Anastasia Ivanova, Jason P. Fine, Additive isotonic proportional hazards models (working in progress).}
\concept{Isotonic regression}
\concept{Survival analysis}
\concept{Nonparametric regression}
\concept{Order-restricted inference}
